[IDB IGNORE] [MDB IGNORE] Merge baymerge-testing into master

+++ code/modules/admin/verbs/SDQL.dm
@@ -1,8 +1,6 @@
 
 //Structured Datum Query Language. Basically SQL meets BYOND objects.
 
-//Note: For use in BS12, need text_starts_with proc, and to modify the action on select to use BS12's object edit command(s).
-
 /client/proc/SDQL_query(query_text as message)
 	set category = "Admin"
 	if(!check_rights(R_DEBUG))  //Shouldn't happen... but just to be safe.
@@ -10,66 +8,66 @@
 
 	var/list/query_list = SDQL_tokenize(query_text)
 
-	if(query_list.len < 2)
-		if(query_list.len > 0)
-			to_chat(usr, "<span class='warning'>SDQL: Too few discrete tokens in query \"[query_text]\". Please check your syntax and try again.</span>")
+	if(length(query_list) < 2)
+		if(length(query_list) > 0)
+			to_chat(usr, SPAN_WARNING("SDQL: Too few discrete tokens in query \"[query_text]\". Please check your syntax and try again."))
 		return
 
 	if(!(lowertext(query_list[1]) in list("select", "delete", "update")))
-		to_chat(usr, "<span class='warning'>SDQL: Unknown query type: \"[query_list[1]]\" in query \"[query_text]\". Please check your syntax and try again.</span>")
+		to_chat(usr, SPAN_WARNING("SDQL: Unknown query type: \"[query_list[1]]\" in query \"[query_text]\". Please check your syntax and try again."))
 		return
 
 	var/list/types = list()
 
 	var/i
-	for(i = 2; i <= query_list.len; i += 2)
+	for(i = 2; i <= length(query_list); i += 2)
 		types += query_list[i]
 
-		if(i + 1 >= query_list.len || query_list[i + 1] != ",")
+		if(i + 1 >= length(query_list) || query_list[i + 1] != ",")
 			break
 
 	i++
 
 	var/list/from = list()
 
-	if(i <= query_list.len)
+	if(i <= length(query_list))
 		if(lowertext(query_list[i]) in list("from", "in"))
-			for(i++; i <= query_list.len; i += 2)
+			for(i++; i <= length(query_list); i += 2)
 				from += query_list[i]
 
-				if(i + 1 >= query_list.len || query_list[i + 1] != ",")
+				if(i + 1 >= length(query_list) || query_list[i + 1] != ",")
 					break
 
 			i++
 
-	if(from.len < 1)
+	if(length(from) < 1)
 		from += "world"
 
 	var/list/set_vars = list()
 
 	if(lowertext(query_list[1]) == "update")
-		if(i <= query_list.len && lowertext(query_list[i]) == "set")
-			for(i++; i <= query_list.len; i++)
-				if(i + 2 <= query_list.len && query_list[i + 1] == "=")
+		if(i <= length(query_list) && lowertext(query_list[i]) == "set")
+			for(i++; i <= length(query_list); i++)
+				if(i + 2 <= length(query_list) && query_list[i + 1] == "=")
 					set_vars += query_list[i]
 					set_vars[query_list[i]] = query_list[i + 2]
 
 				else
-					to_chat(usr, "<span class='warning'>SDQL: Invalid set parameter in query \"[query_text]\". Please check your syntax and try again.</span>")
+					to_chat(usr, SPAN_WARNING("SDQL: Invalid set parameter in query \"[query_text]\". Please check your syntax and try again."))
 					return
 
 				i += 3
 
-				if(i >= query_list.len || query_list[i] != ",")
+				if(i >= length(query_list) || query_list[i] != ",")
 					break
 
-		if(set_vars.len < 1)
-			to_chat(usr, "<span class='warning'>SDQL: Invalid or missing set in query \"[query_text]\". Please check your syntax and try again.</span>")
+		if(length(set_vars) < 1)
+			to_chat(usr, SPAN_WARNING("SDQL: Invalid or missing set in query \"[query_text]\". Please check your syntax and try again."))
 			return
 
 	var/list/where = list()
 
-	if(i <= query_list.len && lowertext(query_list[i]) == "where")
+	if(i <= length(query_list) && lowertext(query_list[i]) == "where")
 		where = query_list.Copy(i + 1)
 
 	var/list/from_objs = list()
@@ -210,15 +208,15 @@
 
 	for(var/datum/t in objs)
 		var/currently_false = 0
-		for(i = 1, i - 1 < where.len, i++)
+		for(i = 1, i - 1 < length(where), i++)
 			var/v = where[i++]
 			var/compare_op = where[i++]
 			if(!(compare_op in list("==", "=", "<>", "<", ">", "<=", ">=", "!=")))
-				to_chat(usr, "<span class='warning'>SDQL: Unknown comparison operator [compare_op] in where clause following [v] in query \"[query_text]\". Please check your syntax and try again.</span>")
+				to_chat(usr, SPAN_WARNING("SDQL: Unknown comparison operator [compare_op] in where clause following [v] in query \"[query_text]\". Please check your syntax and try again."))
 				return
 
 			var/j
-			for(j = i, j <= where.len, j++)
+			for(j = i, j <= length(where), j++)
 				if(lowertext(where[j]) in list("and", "or", ";"))
 					break
 
@@ -246,7 +244,7 @@
 						currently_false = !(value >= result)
 
 
-			if(j > where.len || lowertext(where[j]) == ";")
+			if(j > length(where) || lowertext(where[j]) == ";")
 				break
 			else if(lowertext(where[j]) == "or")
 				if(currently_false)
@@ -261,7 +259,7 @@
 
 
 
-	to_chat(usr, "<span class='notice'>SQDL Query: [query_text]</span>")
+	to_chat(usr, SPAN_NOTICE("SQDL Query: [query_text]"))
 	message_admins("[usr] executed SDQL query: \"[query_text]\".")
 /*
 	for(var/t in types)
@@ -270,7 +268,7 @@
 		to_chat(usr, "From: [t]")
 	for(var/t in set_vars)
 		to_chat(usr, "Set: [t] = [set_vars[t]]")
-	if(where.len)
+	if(length(where))
 		var/where_str = ""
 		for(var/t in where)
 			where_str += "[t] "
@@ -288,7 +286,12 @@
 	switch(lowertext(query_list[1]))
 		if("delete")
 			for(var/datum/t in objs)
-				qdel(t)
+				// turfs are special snowflakes that explode if qdeleted
+				if (isturf(t))
+					var/turf/T = t
+					T.ChangeTurf(world.turf)
+				else
+					qdel(t)
 
 		if("update")
 			for(var/datum/t in objs)
@@ -320,21 +323,13 @@
 					text += "<a href='?src=\ref[t];SDQL_select=\ref[t]'>\ref[t]</a>: [t]<br>"
 
 				//text += "[t]<br>"
-			usr << browse(text, "window=sdql_result")
-
-
-/client/Topic(href,href_list[],hsrc)
-	if(href_list["SDQL_select"])
-		debug_variables(locate(href_list["SDQL_select"]))
-
-	..()
-
+			show_browser(usr, text, "window=sdql_result")
 
 /proc/SDQL_evaluate(datum/object, list/equation)
-	if(equation.len == 0)
+	if(length(equation) == 0)
 		return null
 
-	else if(equation.len == 1)
+	else if(length(equation) == 1)
 		return SDQL_text2value(object, equation[1])
 
 	else if(equation[1] == "!")
@@ -345,7 +340,7 @@
 
 
 	else
-		to_chat(usr, "<span class='warning'>SDQL: Sorry, equations not yet supported :(</span>")
+		to_chat(usr, SPAN_WARNING("SDQL: Sorry, equations not yet supported :("))
 		return null
 
 
@@ -373,16 +368,6 @@
 				return object.vars[text]
 			else
 				return null
-
-
-/proc/text_starts_with(text, start)
-	if(copytext(text, 1, length(start) + 1) == start)
-		return 1
-	else
-		return 0
-
-
-
 
 
 /proc/SDQL_tokenize(query_text)
@@ -430,7 +415,7 @@
 
 		else if(char == "'")
 			if(word != "")
-				to_chat(usr, "<span class=*'warning'>SDQL: You have an error in your SDQL syntax, unexpected ' in query: \"<font color=gray>[query_text]</font>\" following \"<font color=gray>[word]</font>\". Please check your syntax, and try again.</span>")
+				to_chat(usr, SPAN_WARNING("SDQL: You have an error in your SDQL syntax, unexpected ' in query: \"[SPAN_COLOR("gray", query_text)]\" following \"[SPAN_COLOR("gray", word)]\". Please check your syntax, and try again."))
 				return null
 
 			word = "'"
@@ -450,7 +435,7 @@
 					word += char
 
 			if(i > len)
-				to_chat(usr, "<span class='warning'>SDQL: You have an error in your SDQL syntax, unmatched ' in query: \"<font color=gray>[query_text]</font>\". Please check your syntax, and try again.</span>")
+				to_chat(usr, SPAN_WARNING("SDQL: You have an error in your SDQL syntax, unmatched ' in query: \"[SPAN_COLOR("gray", query_text)]\". Please check your syntax, and try again."))
 				return null
 
 			query_list += "[word]'"
@@ -458,7 +443,7 @@
 
 		else if(char == "\"")
 			if(word != "")
-				to_chat(usr, "<span class='warning'>SDQL: You have an error in your SDQL syntax, unexpected \" in query: \"<font color=gray>[query_text]</font>\" following \"<font color=gray>[word]</font>\". Please check your syntax, and try again.</span>")
+				to_chat(usr, SPAN_WARNING("SDQL: You have an error in your SDQL syntax, unexpected \" in query: \"[SPAN_COLOR("gray", query_text)]\" following \"[SPAN_COLOR("gray", word)]\". Please check your syntax, and try again."))
 				return null
 
 			word = "\""
@@ -478,7 +463,7 @@
 					word += char
 
 			if(i > len)
-				to_chat(usr, "<span class='warning'>SDQL: You have an error in your SDQL syntax, unmatched \" in query: \"<font color=gray>[query_text]</font>\". Please check your syntax, and try again.</span>")
+				to_chat(usr, SPAN_WARNING("SDQL: You have an error in your SDQL syntax, unmatched \" in query: \"[SPAN_COLOR("gray", "[query_text]")]\". Please check your syntax, and try again."))
 				return null
 
 			query_list += "[word]\""
