+++ code/modules/integrated_electronics/subtypes/reagents.dm
@@ -1,11 +1,12 @@
 #define IC_SMOKE_REAGENTS_MINIMUM_UNITS 10
 #define IC_REAGENTS_DRAW 0
 #define IC_REAGENTS_INJECT 1
-
+#define IC_HEATER_MODE_HEAT         "heat"
+#define IC_HEATER_MODE_COOL         "cool"
 
 /obj/item/integrated_circuit/reagent
 	category_text = "Reagent"
-	unacidable = 1
+	unacidable = TRUE
 	cooldown_per_use = 10
 	var/volume = 0
 
@@ -116,7 +117,7 @@
 	else
 		direction_mode = IC_REAGENTS_INJECT
 	if(isnum(new_amount))
-		new_amount = Clamp(new_amount, 0, volume)
+		new_amount = clamp(new_amount, 0, volume)
 		transfer_amount = new_amount
 
 
@@ -128,13 +129,13 @@
 			set_pin_data(IC_OUTPUT, 2, weakref(src))
 			push_data()
 
-/obj/item/integrated_circuit/reagent/injector/proc/target_nearby(var/weakref/target)
+/obj/item/integrated_circuit/reagent/injector/proc/target_nearby(weakref/target)
 	var/mob/living/L = target.resolve()
 	if(!L || get_dist(src,L) > 1)
 		return
 	return L
 
-/obj/item/integrated_circuit/reagent/injector/proc/inject_after(var/weakref/target)
+/obj/item/integrated_circuit/reagent/injector/proc/inject_after(weakref/target)
 	busy = FALSE
 	var/mob/living/L = target_nearby(target)
 	if(!L)
@@ -142,12 +143,12 @@
 		return
 	var/atom/movable/acting_object = get_object()
 	log_admin("[key_name(L)] was successfully injected with " + reagents.get_reagents() + " by \the [acting_object]")
-	L.visible_message("<span class='warning'>\The [acting_object] injects [L] with its needle!</span>", \
-					"<span class='warning'>\The [acting_object] injects you with its needle!</span>")
+	L.visible_message(SPAN_WARNING("\The [acting_object] injects [L] with its needle!"), \
+					SPAN_WARNING("\The [acting_object] injects you with its needle!"))
 	reagents.trans_to_mob(L, transfer_amount, CHEM_BLOOD)
 	activate_pin(2)
 
-/obj/item/integrated_circuit/reagent/injector/proc/draw_after(var/weakref/target, var/amount)
+/obj/item/integrated_circuit/reagent/injector/proc/draw_after(weakref/target, amount)
 	busy = FALSE
 	var/mob/living/carbon/C = target_nearby(target)
 	if(!C)
@@ -155,8 +156,8 @@
 		return
 	var/atom/movable/acting_object = get_object()
 
-	C.visible_message("<span class='warning'>\The [acting_object] draws blood from \the [C]</span>",
-					"<span class='warning'>\The [acting_object] draws blood from you.</span>"
+	C.visible_message(SPAN_WARNING("\The [acting_object] draws blood from \the [C]"),
+					SPAN_WARNING("\The [acting_object] draws blood from you.")
 					)
 	C.take_blood(src, amount)
 	activate_pin(2)
@@ -184,15 +185,18 @@
 			var/mob/living/L = AM
 			var/injection_status = L.can_inject(null, BP_CHEST)
 			log_world("Injection status? [injection_status]")
+			var/injection_delay = 3 SECONDS
+			if(injection_status == INJECTION_PORT)
+				injection_delay += INJECTION_PORT_DELAY
 			if(!injection_status)
 				activate_pin(3)
 				return
 			//Always log attemped injections for admins
 			log_admin("[key_name(L)] is getting injected with " + reagents.get_reagents() + " by \the [acting_object]")
-			L.visible_message("<span class='danger'>\The [acting_object] is trying to inject [L]!</span>", \
-								"<span class='danger'>\The [acting_object] is trying to inject you!</span>")
+			L.visible_message(SPAN_DANGER("\The [acting_object] is trying to inject [L]!"), \
+								SPAN_DANGER("\The [acting_object] is trying to inject you!"))
 			busy = TRUE
-			addtimer(CALLBACK(src, .proc/inject_after, weakref(L)), injection_status * 3 SECONDS)
+			addtimer(new Callback(src, .proc/inject_after, weakref(L)), injection_delay)
 			return
 		else
 			if(!AM.is_open_container())
@@ -202,7 +206,7 @@
 
 			reagents.trans_to(AM, transfer_amount)
 
-	if(direction_mode == IC_REAGENTS_DRAW)
+	else if(direction_mode == IC_REAGENTS_DRAW)
 		if(reagents.total_volume >= reagents.maximum_volume)
 			acting_object.visible_message("\The [acting_object] tries to draw from [AM], but the injector is full.")
 			activate_pin(3)
@@ -210,21 +214,24 @@
 
 		var/tramount = abs(transfer_amount)
 
-		if(ishuman(AM))
-			var/mob/living/carbon/human/H = AM
-			var/injection_status = H.can_inject(null, BP_CHEST)
-			if(!H.dna || !injection_status || !H.vessel.total_volume)
+		if(istype(AM, /mob/living/carbon))
+			var/mob/living/carbon/C = AM
+			var/injection_status = C.can_inject(null, BP_CHEST)
+			var/injection_delay = 3 SECONDS
+			if(injection_status == INJECTION_PORT)
+				injection_delay += INJECTION_PORT_DELAY
+			if(istype(C, /mob/living/carbon/slime) || !C.dna || !injection_status)
 				activate_pin(3)
 				return
-			H.visible_message("<span class='danger'>\The [acting_object] is trying to take a blood sample from [H]!</span>", \
-								"<span class='danger'>\The [acting_object] is trying to take a blood sample from you!</span>")
+			C.visible_message(SPAN_DANGER("\The [acting_object] is trying to take a blood sample from [C]!"), \
+								SPAN_DANGER("\The [acting_object] is trying to take a blood sample from you!"))
 			busy = TRUE
-			addtimer(CALLBACK(src, .proc/draw_after, weakref(H), tramount), injection_status * 3 SECONDS)
+			addtimer(new Callback(src, .proc/draw_after, weakref(C), tramount), injection_delay)
 			return
 
 		else
 			if(!AM.reagents.total_volume)
-				acting_object.visible_message("<span class='notice'>\The [acting_object] tries to draw from [AM], but it is empty!</span>")
+				acting_object.visible_message(SPAN_NOTICE("\The [acting_object] tries to draw from [AM], but it is empty!"))
 				activate_pin(3)
 				return
 
@@ -263,7 +270,7 @@
 	else
 		direction_mode = IC_REAGENTS_INJECT
 	if(isnum(new_amount))
-		new_amount = Clamp(new_amount, 0, 50)
+		new_amount = clamp(new_amount, 0, 50)
 		transfer_amount = new_amount
 
 /obj/item/integrated_circuit/reagent/pump/do_work()
@@ -305,7 +312,7 @@
 		"volume used" = IC_PINTYPE_NUMBER,
 		"self reference" = IC_PINTYPE_REF
 		)
-	activators = list("push ref" = IC_PINTYPE_PULSE_OUT)
+	activators = list("push ref" = IC_PINTYPE_PULSE_IN)
 	spawn_flags = IC_SPAWN_DEFAULT|IC_SPAWN_RESEARCH
 
 
@@ -374,12 +381,18 @@
 		activate_pin(3)
 		return FALSE
 	var/obj/item/I = get_pin_data_as_type(IC_INPUT, 1, /obj/item)
+
+	if(isnull(I))
+		return FALSE
+
 	if(!I.reagents || !I.reagents.total_volume)
 		activate_pin(3)
 		return FALSE
+
 	I.reagents.trans_to(src,I.reagents.total_volume)
 	if(!I.reagents.total_volume)
 		qdel(I)
+
 	activate_pin(2)
 	return FALSE
 
@@ -420,8 +433,8 @@
 	desc = "Filters liquids by list of desired or unwanted reagents."
 	icon_state = "reagent_filter"
 	extended_desc = "This is a filter which will move liquids from the source to its target. \
-	If the amount in the fourth pin is positive, it will move all reagents except those in the unwanted list. \
-	If the amount in the fourth pin is negative, it will only move the reagents in the wanted list. \
+	If the amount in the third pin is positive, it will move all reagents except those in the unwanted list. \
+	If the amount in the third pin is negative, it will only move the reagents in the wanted list. \
 	The third pin determines how many reagents are moved per pulse, between 0 and 50. Amount is given for each separate reagent."
 
 	complexity = 8
@@ -452,7 +465,7 @@
 	else
 		direction_mode = IC_REAGENTS_INJECT
 	if(isnum(new_amount))
-		new_amount = Clamp(new_amount, 0, 50)
+		new_amount = clamp(new_amount, 0, 50)
 		transfer_amount = new_amount
 
 /obj/item/integrated_circuit/reagent/filter/do_work()
@@ -483,5 +496,153 @@
 	activate_pin(2)
 	push_data()
 
+// This is an input circuit because attackby_react is only called for input circuits
+/obj/item/integrated_circuit/input/funnel
+	category_text = "Reagent"
+	name = "reagent funnel"
+	desc = "A funnel with a small pump that lets you refill an internal reagent storage."
+	icon_state = "reagent_funnel"
+
+	inputs = list(
+		"target" = IC_PINTYPE_REF
+	)
+	activators = list(
+		"on transfer" = IC_PINTYPE_PULSE_OUT
+	)
+
+	unacidable = TRUE
+	spawn_flags = IC_SPAWN_DEFAULT|IC_SPAWN_RESEARCH
+	complexity = 4
+	power_draw_per_use = 5
+
+/obj/item/integrated_circuit/input/funnel/attackby_react(obj/item/I, mob/living/user, intent)
+	var/atom/movable/target = get_pin_data_as_type(IC_INPUT, 1, /atom/movable)
+	var/obj/item/reagent_containers/container = I
+
+	if(!check_target(target))
+		return FALSE
+
+	if(!istype(container))
+		return FALSE
+
+	// Messages are provided by standard_pour_into
+	if(container.standard_pour_into(user, target))
+		activate_pin(1)
+		return TRUE
+
+	return FALSE
+
+// Most of this is just chemical heater code refitted for ICs
+/obj/item/integrated_circuit/reagent/temp
+	inputs = list(
+		"target temperature" = IC_PINTYPE_NUMBER
+	)
+	outputs = list(
+		"volume used" = IC_PINTYPE_NUMBER,
+		"temperature" = IC_PINTYPE_NUMBER,
+		"enabled" = IC_PINTYPE_BOOLEAN,
+		"self reference" = IC_PINTYPE_REF
+	)
+	activators = list(
+		"toggle" = IC_PINTYPE_PULSE_IN,
+		"on toggle" = IC_PINTYPE_PULSE_OUT,
+		"push ref" = IC_PINTYPE_PULSE_IN
+	)
+
+	atom_flags = ATOM_FLAG_OPEN_CONTAINER
+	complexity = 12
+	cooldown_per_use = 1
+	power_draw_per_use = 50
+	volume = 30
+
+	var/active = 0
+	var/min_temp = 40 CELSIUS
+	var/max_temp = 200 CELSIUS
+	var/heating_power = 5
+	var/target_temp = T20C
+	var/last_temperature = 0
+	var/mode = IC_HEATER_MODE_HEAT
+
+/obj/item/integrated_circuit/reagent/temp/Initialize()
+	. = ..()
+
+	set_pin_data(IC_OUTPUT, 2, temperature - T0C)
+	push_data()
+
+/obj/item/integrated_circuit/reagent/temp/do_work(ord)
+	switch(ord)
+		if(1)
+			target_temp = get_pin_data(IC_INPUT, 1)
+			if(isnull(target_temp))
+				return
+
+			// +/- T0C to convert to/from kelvin
+			target_temp = clamp(target_temp + T0C, min_temp, max_temp)
+			set_pin_data(IC_INPUT, 1, target_temp - T0C)
+
+			active = !active
+			set_pin_data(IC_OUTPUT, 3, active)
+			push_data()
+			activate_pin(2)
+
+			// begin processing temperature
+			if(active)
+				QUEUE_TEMPERATURE_ATOMS(src)
+		if(3)
+			set_pin_data(IC_OUTPUT, 4, weakref(src))
+			push_data()
+
+/obj/item/integrated_circuit/reagent/temp/on_reagent_change()
+	push_vol()
+
+/obj/item/integrated_circuit/reagent/temp/power_fail()
+	active = 0
+
+/obj/item/integrated_circuit/reagent/temp/ProcessAtomTemperature()
+	if(!active)
+		return PROCESS_KILL
+
+	last_temperature = temperature
+
+	if(mode == IC_HEATER_MODE_HEAT && temperature < target_temp)
+		temperature = min(temperature + heating_power, max_temp)
+	else if(mode == IC_HEATER_MODE_COOL && temperature > target_temp)
+		temperature = max(temperature - heating_power, min_temp)
+
+	if(temperature != last_temperature)
+		// Lost power
+		if(!check_power())
+			power_fail()
+			return ..()
+
+		set_pin_data(IC_OUTPUT, 2, temperature - T0C)
+		push_data()
+
+	return TRUE
+
+/obj/item/integrated_circuit/reagent/temp/heater
+	name = "reagent heater"
+	desc = "A small reagent container capable of heating reagents. It can hold up to 30u."
+	icon_state = "reagent_heater"
+	extended_desc = "This is effectively an internal beaker. It has a heating coil wrapped around it, which allows it to heat the contents of the beaker. Temperature is given in celsius."
+
+	spawn_flags = IC_SPAWN_RESEARCH
+
+/obj/item/integrated_circuit/reagent/temp/cooler
+	name = "reagent cooler"
+	desc = "A small reagent container capable of cooling reagents. It can hold up to 30u."
+	icon_state = "reagent_cooler"
+	extended_desc = "This is effectively an internal beaker. It has a cooling mechanism wrapped around it, which allows it to cool the contents of the beaker. Temperature is given in celsius."
+
+	spawn_flags = IC_SPAWN_RESEARCH
+
+	min_temp = -80 CELSIUS
+	max_temp = 30 CELSIUS
+	mode = IC_HEATER_MODE_COOL
+
+//undefs
+
+#undef IC_HEATER_MODE_HEAT
+#undef IC_HEATER_MODE_COOL
 #undef IC_REAGENTS_DRAW
 #undef IC_REAGENTS_INJECT
